rename jiracli to jira_cmd
