+++ jira_cmd/commands/issue/close.py
@@ -5,10 +5,10 @@
 Command module for creating Jira issues.
 """
 
-from jiracli.commands.issue import JiraIssueCommand
+from jira_cmd.commands.issue import JiraIssueCommand
 
 
-class IssueAttach(JiraIssueCommand):
+class IssueClose(JiraIssueCommand):
     """
     Command class for creating Jira issues.
     """
@@ -21,7 +21,16 @@ class IssueAttach(JiraIssueCommand):
             parser: The argument parser to add arguments to
         """
         super().define_arguments(parser)
+        parser.add_argument(
+            "-i",
+            "--issue",
+            help="Issue to match",
+            default=None,
+            required=True,
+        )
         return parser
 
     def execute(self, args):
-        pass
+        issue = self.jira_client.issue(args.issue)
+        issue.transition("Close")
+        self.log.info("Issue %s closed", args.issue)
