+++ jira_cmd/commands/issue/label.py
@@ -1,9 +1,10 @@
 # #!/usr/bin/env python
 # pylint: disable=no-member
 
-from jiracli.commands.issue import JiraIssueCommand
+from jira_cmd.commands.issue import JiraIssueCommand
 
-class IssueComment(JiraIssueCommand):
+
+class IssueLabel(JiraIssueCommand):
     """
     Command class for commenting on Jira issues.
     """
@@ -25,9 +26,9 @@ class IssueComment(JiraIssueCommand):
             required=True,
         )
         parser.add_argument(
-            "-c",
-            "--comment",
-            help="Comment to add",
+            "-l",
+            "--labels",
+            help="Labels to add",
             default=None,
             required=True,
         )
@@ -40,6 +41,6 @@ class IssueComment(JiraIssueCommand):
         Args:
             args: The parsed arguments
         """
-        self.log.debug("Commenting on Jira issue")
-        self.jira_client.add_comment(args.issue, args.comment)
-        self.log.info("Added comment to %s", args.issue)
+        self.log.debug("Adding lables to Jira issue %s", args.issue)
+        self.jira_client.add_labels(args.issue, args.labels)
+        self.log.info("Issue %s labels updated", args.issue)
